Refactor database connection handling: replace dbConfig with mongoConnection and clean up code formatting
